moodel_detection/requirements.txt
tensorflow
Pillow
Flask
gunicorn